# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Order API Documentation",
        "description": "API documentation for your Food Ordering Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/admin/orders": {
            "get": {
                "tags": [
                    "Orders (Admin)"
                ],
                "summary": "Get all orders",
                "operationId": "c2284ff5af1b95581dc4bcf924f4c0e3",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "Filter by orders id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders (Admin)"
                ],
                "summary": "Create new orders",
                "operationId": "6097de4b0ce7747492d384048192a4b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "ordertype_id",
                                    "paymenttype_id",
                                    "stage_id",
                                    "subtotal",
                                    "discount",
                                    "delivery_fee",
                                    "service_fee",
                                    "total",
                                    "is_paid"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ordertype_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "paymenttype_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stage_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    },
                                    "delivery_fee": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3
                                    },
                                    "service_fee": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "transaction_id": {
                                        "type": "string",
                                        "example": "TX123456789"
                                    },
                                    "is_paid": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "order_note": {
                                        "type": "string",
                                        "example": "Please deliver between 5-6 PM"
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-15T17:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Orders (Admin)"
                ],
                "summary": "Get order by ID",
                "operationId": "7fd828033c0c6a990484d7070919b15d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/orders/{id}/stage": {
            "patch": {
                "tags": [
                    "Orders (Admin)"
                ],
                "summary": "Update order",
                "operationId": "25d2ac528dc34b1b77dc7b7020354738",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "order id to update stage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stage_id",
                        "in": "path",
                        "description": "update order stage id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}